Moved a file to appropriate folder
